Update README.md
Hi there, thanks for your project to let us have a chance to test firebase project with cypress.
You have done a great innovation job! I will continually follow your new footage. 

For this PR, I am just want to mention something about ```firebase-tools``` installation.

The case is that I failed to perform DELETE to Firestore using your ```cy.callFirestore```, and I found the error message in cypress is something like we haven’t ```firebase``` in ```node_modules/.bin```, we know that this folder contains some of the files as a link(symbol) of our packages installed in project(automatically) or globally(maybe we need do some extra job). 

Then after my observation, I found the ```firebase``` symbol is related to ```firebase-tools```, so each time we perform DELETE, we need this file in our ```node_modules/.bin```.

After that, I tried to reinstall  ```firebase-tools``` again globally, and I still can’t see that symbol created in that folder.

Then I tried to install it locally(in the project), and it appeared, and finally problem solved.

So I suggest to change the document to guide the user to install ```firebase-tools``` in project and globally.
I think it has no harm to this in both ways.

Cheers.