Update README.md
`callFirestore("get", ...)` appears to return a single instance, not an array of instances.  `r[0]` is undefined.  `r.age` and `r.name` exist.  There does not appear to be a clear way to get `r.id` (though the caller knows the ID since they used it in the "get" call . . .)

If there is a way to get the metadata from a Firestore doc, would be GREAT to add that to the example.